HACK: Comment out separate destruction of zink_batch_states in context
The zink_batch_state object is destroyed twice as it is stored directly
in the context and in a list of batch states in the context.

Being destroyed twice seems to be some symptom exposed only when running
with Juice and/or crashing in the Vulkan library (as is happening for
Juice).  This isn't a problem with a native run.