FIX: Populate swapchain preTransform from surface capabilities
Except this is not working because, I think, I'm creating the surface
against the dummy window created to initialize the actual OpenGL
context that comes later.  This means that that window and surface have
been destroyed by the time the true window is being initialized.

The fix is to move the surface from the screen to the context so that
it is created properly.