WIP: Add zink_bo_wrap_res() to create a zink_bo wrapping a VkImage
Need to check that destroying the zink_bo_wrap_res() doesn't destroy
the image or try to reuse any memory that Zink and/or Mesa think it
might be using.

This was added to wrap VkImages returned from the swapchain in Zink and
Mesa compatible resources.  Based on d3d12_bo_wrap_res().