WIP: Signal and wait on draw and present finished semaphores
Signals the draw finished semaphore on the end of frame flush submit.
Waits on the present finished semaphore on first batch in the following
frame.

FIX: Set pipeline stage to wait on before drawing to swapchain image

WIP: Flush even empty batches for the end of frame flush

This is so that the semaphore that signals presentation of an image
has finished is always waited on.

FIX: Add present finished to batch state's wait semaphores in flush_batch()

This happens on the first batch submitted after the swapchain has been
presented.

FIX: Include <GL/gl.h> from zink_context.c (hack to access zink_wgl_framebuffer)