WIP: Set vertex binding stride without EXT_vertex_input_dynamic_state
When EXT_vertex_input_dynamic_state is not available stride is never
set.  This copies stride into the VkVertexInputBindingDescription
array just before the graphics pipeline is created.

I believe this is compatible with hashing because the vertex_strides
array is included as part of the graphics pipeline hash shortly before
zink_create_gfx_pipeline() is called.