WIP: zink: Always enable DYNAMIC_STATE_VERTEX_INPUT_BINDING_STRIDE_EXT
Previously this was enabled only if EXT_vertex_input_dynamic_state
wasn't used and was generating validation layer warnings as stride
was being set dynamically.

I'm not certain that the better fix might be to stop setting stride
dynamically when have_EXT_vertex_input_dynamic_state is available but
this was the fix I found first.