HACK: Comment assert PIPE_BIND_DISPLAY_TARGET in zink_kopper_fixup_depth_buffer()
I'm pretty sure that this assert is not needed?  The
PIPE_BIND_DISPLAY_TARGET flag indicates that the buffer is... a front
buffer or allocated by the windowing system.  The depth buffer will
never be a front buffer, possibly it could be considered allocated
by the windowing system in some logical way.  I'll comment it out
for now and see if other seemingly related things break.