WIP: Hold reference to zink_context::needs_present to avoid dangling it
The resource pointed to be needs_present appears to be destroyed leaving
needs_present dangling.  There are lots of crashing and asserts that
I can trace back to needs_present being invalid and often the memory
pointed to is 0xdddddddd which looks like MSVC sentinel values for
deleted or destroyed memory.