Track resources in usage context not current context
When a resource is unbound its batch usage is checked to see if the
batch needs to additionally track the resource.  This tracking is so
that usage can be reset correctly, especially when the context is
destroyed.  The resource's write usage is reapplied in the current
context which also adds tracking to the current context.  The read
usage is left, presumably because reapplying the write usage is enough
to delay the resource's resource or destruction until after the GPU has
finished with the resource.

However, in the presence of multiple contexts, the resource's read and
write usage can refer to a batch from another context that isn't the
current context.  When this other context is destroyed before the
current context this can leave the read usage dangling to a batch in
the other context and this leads to a crash when that resource's usage
is checked again.

Fixed by restoring tracking to the context that owns the usage rather
to the current context.  This has the resources tracked from the
correct contexts and avoids leaving dangling usages.

Another solution is to reapply both the read and write usages in the
current context but that seems less reliable because there doesn't
seem any guarantee that the read or write usage from one context will
be complete before the reapplied usage in another context.